Feature back test guppy mma strategy

+++ TraderDashboardUi.Entity/PositionSize.cs
@@ -6,11 +6,11 @@ namespace TraderDashboardUi.Entity
 {
     public class PositionSize
     {
-        public double RiskLimit { get; set; }
-        public double AccountSize { get; set; }
-        public double PipRisk { get; set; }
-        public double PipValue { get; set; }
-        public double Size { get; set; }
+        public decimal RiskLimit { get; set; }
+        public decimal AccountSize { get; set; }
+        public decimal PipRisk { get; set; }
+        public decimal PipValue { get; set; }
+        public decimal Size { get; set; }
 
         // pipRisk * pipValue * lots traded = amount of risk
     }
